+++ src/stage_node.cpp
@@ -190,68 +190,6 @@ int StageNode::callback_init_stage_model(Stg::Model * mod, StageNode * node)
   return 0;
 }
 
-geometry_msgs::msg::Quaternion StageNode::createQuaternionMsgFromYaw(double yaw)
-{
-  tf2::Quaternion q;
-  q.setRPY(0, 0, yaw);
-  return tf2::toMsg(q);
-}
-
-void StageNode::publish_object_visualization(StageNode * node)
-{
-  rclcpp::QoS qos(10);
-  rclcpp::Publisher<visualization_msgs::msg::MarkerArray>::SharedPtr pub_object = node->create_publisher<visualization_msgs::msg::MarkerArray>("obstacle_viszualization", qos);
-  visualization_msgs::msg::MarkerArray marker_array;
-
-  for (auto object: node->objects_) {
-
-    visualization_msgs::msg::Marker marker_text;
-    marker_text.id = object->id();
-    marker_text.header.frame_id = "map";
-    marker_text.header.stamp = object->node()->sim_time_;
-    marker_text.pose.position.x = object->model->GetGlobalPose().x;
-    marker_text.pose.position.y = object->model->GetGlobalPose().y;
-    marker_text.pose.position.z = object->model->GetGlobalPose().z;
-    marker_text.pose.orientation = createQuaternionMsgFromYaw(object->model->GetGlobalPose().a);
-    marker_text.type = visualization_msgs::msg::Marker::TEXT_VIEW_FACING;
-    marker_text.text = object->name();
-    marker_text.ns = "ObjectNames";
-    marker_text.action = visualization_msgs::msg::Marker::MODIFY;
-    marker_text.scale.z = 0.3;
-    marker_text.color.r = 1.0;
-    marker_text.color.g = 1.0;
-    marker_text.color.b = 1.0;
-    marker_text.color.a = 1.0;
-
-
-    visualization_msgs::msg::Marker marker_pose;
-    marker_pose.id = object->id();
-    marker_pose.header.frame_id = "map";
-    marker_pose.header.stamp = object->node()->sim_time_;
-    marker_pose.pose.position.x = object->model->GetGlobalPose().x;
-    marker_pose.pose.position.y = object->model->GetGlobalPose().y;
-    marker_pose.pose.position.z = object->model->GetGlobalPose().z;
-    marker_pose.pose.orientation = createQuaternionMsgFromYaw(object->model->GetGlobalPose().a);
-    marker_pose.type = visualization_msgs::msg::Marker::CUBE;
-    marker_pose.ns = "ObjectPoses";
-    marker_pose.action = visualization_msgs::msg::Marker::MODIFY;
-    marker_pose.scale.x = object->model->GetGeom().size.x;
-    marker_pose.scale.y = object->model->GetGeom().size.y;
-    marker_pose.scale.z = object->model->GetGeom().size.z;
-    marker_pose.color.r = 255.0;
-    marker_pose.color.g = 0.0;
-    marker_pose.color.b = 0.0;
-    marker_pose.color.a = 1.0;
-
-    marker_array.markers.push_back(marker_text);
-    marker_array.markers.push_back(marker_pose);
-
-  }
-  
-  pub_object->publish(marker_array);
-
-}
-
 int StageNode::callback_update_stage_world(Stg::World * world, StageNode * node)
 {
   // We return false to indicate that we want to be called again (an
@@ -310,7 +248,7 @@ bool StageNode::cb_reset_srv(
   for (auto vehicle: this->vehicles_) {
     vehicle->soft_reset();
   }
-  for (auto object: this->objects_) {
+  for (const auto& object: this->objects_) {
     object->soft_reset();
   }
   return true;
@@ -321,7 +259,7 @@ void StageNode::cb_object_setpose_srv(
   std::shared_ptr<stage_ros2::srv::SetObjectPose::Response> response)
 {
   RCLCPP_INFO(this->get_logger(), "Setting Position of %s!", request->name.c_str());
-  for (auto object: this->objects_) {
+  for (const auto& object: this->objects_) {
     if(object->name() == request->name){
       Stg::Pose pose = Stg::Pose(request->x, request->y, request->z, request->yaw);
       object->model->SetPose(pose);
@@ -343,7 +281,7 @@ void StageNode::cb_object_setpose_from_robot_srv(
   double y = vehicle->positionmodel->GetGlobalPose().y + std::sin(vehicle->positionmodel->GetGlobalPose().a) * request->x + std::cos(vehicle->positionmodel->GetGlobalPose().a) * request->y;
   double yaw = vehicle->positionmodel->GetGlobalPose().a + request->yaw;
 
-  for (auto object: this->objects_) {
+  for (const auto& object: this->objects_) {
     if(object->name() == request->name){
       Stg::Pose pose = Stg::Pose(x, y, 0, yaw);
       object->model->SetPose(pose);
@@ -389,7 +327,7 @@ int StageNode::SubscribeModels()
     vehicle->init(use_topic_prefix, this->one_tf_tree_);
   }
 
-  for (std::shared_ptr<Object> object: this->objects_) {
+  for (const auto& object: this->objects_) {
     object->init();
   }
 
@@ -436,4 +374,11 @@ geometry_msgs::msg::TransformStamped StageNode::create_transform_stamped(
   out.transform.rotation.y = in.getRotation().getY();
   out.transform.rotation.z = in.getRotation().getZ();
   return out;
+}
+
+geometry_msgs::msg::Quaternion StageNode::createQuaternionMsgFromYaw(double yaw)
+{
+  tf2::Quaternion q;
+  q.setRPY(0, 0, yaw);
+  return tf2::toMsg(q);
 }
